back/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
axios==0.4.0
beautifulsoup4==4.12.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==41.0.3
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib_metadata==7.1.0
ipykernel==6.29.1
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.6.0
oauthlib==3.2.2
openai==1.7.0
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
selenium==4.19.0
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.13
wsproto==1.2.0
zipp==3.18.1
